Change search strategy to an enum and split code out to module
