+++ Dockerfile
@@ -1,6 +1,7 @@
 FROM python:3.12-bookworm
 
 COPY entrypoint.py /entrypoint.py
+COPY cofactr_cogs /cofactr_cogs
 
 RUN pip install requests
